requirements.txt
alembic==1.10.3
anyio==3.7.1
bardapi==0.1.29
beautifulsoup4==4.12.2
beepy==1.0.7
browser-cookie3==0.19.1
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.2.0
deep-translator==1.11.4
dominate==2.7.0
exceptiongroup==1.1.2
google-ai-generativelanguage==0.2.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.56.2
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
lz4==4.3.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.24.3
pandas==2.0.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycryptodomex==3.18.0
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytz==2023.3
pywin32==306
requests==2.28.2
rsa==4.9
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.19
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
zipp==3.15.0
